+++ sqlitefile/buffer_test.go
@@ -108,7 +108,9 @@ func TestBuffer(t *testing.T) {
 func TestConcurrentBuffer(t *testing.T) {
 	// Make sure the shared cache table lock does not
 	// apply to blob buffers (because we use temp tables).
-	dbpool, err := sqlitex.Open("file::memory:?mode=memory", 0, 2)
+	dbpool, err := sqlitex.NewPool("file::memory:?mode=memory", sqlitex.PoolOptions{
+		PoolSize: 2,
+	})
 	if err != nil {
 		t.Fatal(err)
 	}